Removes ssh_opt test
Due to refactoring, the CNAMES are not created with the build.sh script, so the CNAMES
won't be available, and you won't be able to lookup by name, which is what that test is trying
to do.

TODO: Need to add a test for running the load balancer setup script, and then re-add the ssh_opt
test -- either after the load balancer setup script, or figure out a way to run it without requiring
the --name param.