Allows using existing key and group
When need to be able to build new machines that can be used as replacements for
instances in an existing demo/prod pair, but this is tricky because the 'name'
parameter is so pervasive as a convention in tying all the AWS resources
together. This PR adds a couple of flags that allow a build to skip the
creation of a Key pair and an IAM group, assuming that it can use a Key pair
and group that already exist, based on the given 'name' parameter.

Closes #89.